requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
idna==3.4
loguru==0.6.0
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
requests==2.28.2
sqlparse==0.4.3
telebot==0.0.5
urllib3==1.26.14
whatsapp-api-client-python==0.0.31
djangorestframework==3.14.0
django-environ==0.10.0